requirements.txt
APScheduler==3.7.0
Flask==2.0.1
gunicorn
Jinja2==3.0.1
lxml==4.6.3
Pillow==8.3.1
requests==2.22.0
urllib3==1.25.8
pandas==1.2.5